refactor: extract runtime aux-wiring to a sibling to hold the module-size budget
